tests/submission/W3C/domhtml/HTMLAnchorElement11.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - Anchor</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/the-a-element.html#dom-a-target" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<p>
<a href="/support/images/yellow-rectangle.png" target="dynamic">View Submit Button</a>
</p>


<script>
var domtest = async_test("The target attribute specifies the frame to render the source in.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vtarget;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("a");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vtarget = testNode.target;

      assert_equals(vtarget,"dynamic");
       
    domtest.done();
}));

</script>
<p>The target attribute specifies the frame to render the source in.

    Retrieve the target attribute and examine it's value.</p>
<div id="log"></div>
</body></html>
